+++ FLAlgorithms/servers/serverpFedMe.py
@@ -1,28 +1,31 @@
 import torch
 import os
 
-from fedl.users.userperavg import UserPerAvg
-from fedl.servers.serverbase import Server
+from FLAlgorithms.users.userpFedMe import UserpFedMe
+from FLAlgorithms.servers.serverbase import Server
 from utils.model_utils import read_data, read_user_data
+import numpy as np
+ 
+# Implementation for pFedMe Server
 
-# Implementation for per-FedAvg Server
-
-class PerAvg(Server):
-    def __init__(self, dataset,algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters,
-                 local_epochs, optimizer, num_users,times):
+class pFedMe(Server):
+    def __init__(self, dataset, algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters,
+                 local_epochs, optimizer, num_users, K, personal_learning_rate, times):
         super().__init__(dataset,algorithm, model[0], batch_size, learning_rate, beta, lamda, num_glob_iters,
                          local_epochs, optimizer, num_users, times)
 
         # Initialize data for all  users
         data = read_data(dataset)
         total_users = len(data[0])
+        self.K = K
+        self.personal_learning_rate = personal_learning_rate
         for i in range(total_users):
             id, train , test = read_user_data(i, data, dataset)
-            user = UserPerAvg(id, train, test, model, batch_size, learning_rate, beta, lamda, local_epochs, optimizer ,total_users , num_users)
+            user = UserpFedMe(id, train, test, model, batch_size, learning_rate, beta, lamda, local_epochs, optimizer, K, personal_learning_rate)
             self.users.append(user)
             self.total_train_samples += user.train_samples
         print("Number of users / total users:",num_users, " / " ,total_users)
-        print("Finished creating Local Per-Avg.")
+        print("Finished creating pFedMe server.")
 
     def send_grads(self):
         assert (self.users is not None and len(self.users) > 0)
@@ -43,16 +46,28 @@ class PerAvg(Server):
             self.send_parameters()
 
             # Evaluate gloal model on user for each interation
-            print("Evaluate global model with one step update")
+            print("Evaluate global model")
             print("")
-            self.evaluate_one_step()
+            self.evaluate()
 
-            # choose several users to send back upated model to server
-            self.selected_users = self.select_users(glob_iter,self.num_users)
-            for user in self.selected_users:
+            # do update for all users not only selected users
+            for user in self.users:
                 user.train(self.local_epochs) #* user.train_samples
-                
-            self.aggregate_parameters()
+            
+            # choose several users to send back upated model to server
+            # self.personalized_evaluate()
+            self.selected_users = self.select_users(glob_iter,self.num_users)
 
+            # Evaluate gloal model on user for each interation
+            #print("Evaluate persionalized model")
+            #print("")
+            self.evaluate_personalized_model()
+            #self.aggregate_parameters()
+            self.persionalized_aggregate_parameters()
+
+
+        #print(loss)
         self.save_results()
         self.save_model()
+    
+
