+++ FLAlgorithms/servers/serverperavg.py
@@ -1,31 +1,28 @@
 import torch
 import os
 
-from fedl.users.userpsnl import UserpFedMe
-from fedl.servers.serverbase import Server
+from FLAlgorithms.users.userperavg import UserPerAvg
+from FLAlgorithms.servers.serverbase import Server
 from utils.model_utils import read_data, read_user_data
-import numpy as np
- 
-# Implementation for pFedMe Server
 
-class pFedMe(Server):
-    def __init__(self, dataset, algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters,
-                 local_epochs, optimizer, num_users, K, personal_learning_rate, times):
+# Implementation for per-FedAvg Server
+
+class PerAvg(Server):
+    def __init__(self, dataset,algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters,
+                 local_epochs, optimizer, num_users,times):
         super().__init__(dataset,algorithm, model[0], batch_size, learning_rate, beta, lamda, num_glob_iters,
                          local_epochs, optimizer, num_users, times)
 
         # Initialize data for all  users
         data = read_data(dataset)
         total_users = len(data[0])
-        self.K = K
-        self.personal_learning_rate = personal_learning_rate
         for i in range(total_users):
             id, train , test = read_user_data(i, data, dataset)
-            user = UserpFedMe(id, train, test, model, batch_size, learning_rate, beta, lamda, local_epochs, optimizer, K, personal_learning_rate)
+            user = UserPerAvg(id, train, test, model, batch_size, learning_rate, beta, lamda, local_epochs, optimizer ,total_users , num_users)
             self.users.append(user)
             self.total_train_samples += user.train_samples
         print("Number of users / total users:",num_users, " / " ,total_users)
-        print("Finished creating pFedMe server.")
+        print("Finished creating Local Per-Avg.")
 
     def send_grads(self):
         assert (self.users is not None and len(self.users) > 0)
@@ -46,28 +43,16 @@ class pFedMe(Server):
             self.send_parameters()
 
             # Evaluate gloal model on user for each interation
-            print("Evaluate global model")
+            print("Evaluate global model with one step update")
             print("")
-            self.evaluate()
+            self.evaluate_one_step()
 
-            # do update for all users not only selected users
-            for user in self.users:
-                user.train(self.local_epochs) #* user.train_samples
-            
             # choose several users to send back upated model to server
-            # self.personalized_evaluate()
             self.selected_users = self.select_users(glob_iter,self.num_users)
+            for user in self.selected_users:
+                user.train(self.local_epochs) #* user.train_samples
+                
+            self.aggregate_parameters()
 
-            # Evaluate gloal model on user for each interation
-            #print("Evaluate persionalized model")
-            #print("")
-            self.evaluate_personalized_model()
-            #self.aggregate_parameters()
-            self.persionalized_aggregate_parameters()
-
-
-        #print(loss)
         self.save_results()
         self.save_model()
-    
-
