--- fedl/users/userperavg.py
@@ -1,78 +1,0 @@
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-import os
-import json
-from torch.utils.data import DataLoader
-from fedl.optimizers.fedoptimizer import MySGD, FEDLOptimizer
-from fedl.users.userbase import User
-
-# Implementation for Per-FedAvg clients
-
-class UserPerAvg(User):
-    def __init__(self, numeric_id, train_data, test_data, model, batch_size, learning_rate,beta,lamda,
-                 local_epochs, optimizer, total_users , num_users):
-        super().__init__(numeric_id, train_data, test_data, model[0], batch_size, learning_rate, beta, lamda,
-                         local_epochs)
-        self.total_users = total_users
-        self.num_users = num_users
-        
-        if(model[1] == "Mclr_CrossEntropy"):
-            self.loss = nn.CrossEntropyLoss()
-        else:
-            self.loss = nn.NLLLoss()
-
-        self.optimizer = MySGD(self.model.parameters(), lr=self.learning_rate)
-
-    def set_grads(self, new_grads):
-        if isinstance(new_grads, nn.Parameter):
-            for model_grad, new_grad in zip(self.model.parameters(), new_grads):
-                model_grad.data = new_grad.data
-        elif isinstance(new_grads, list):
-            for idx, model_grad in enumerate(self.model.parameters()):
-                model_grad.data = new_grads[idx]
-
-    def train(self, epochs):
-        LOSS = 0
-        self.model.train()
-        for epoch in range(1, self.local_epochs + 1):  # local update 
-            
-            self.model.train()
-
-            #step 1
-            X, y = self.get_next_train_batch()
-            self.optimizer.zero_grad()
-            output = self.model(X)
-            loss = self.loss(output, y)
-            loss.backward()
-            self.optimizer.step()
-
-            #step 2
-            X, y = self.get_next_train_batch()
-            self.optimizer.zero_grad()
-            output = self.model(X)
-            loss = self.loss(output, y)
-            loss.backward()
-            self.optimizer.step(beta = self.beta)
-
-            # clone model to user model 
-            self.clone_model_paramenter(self.model.parameters(), self.local_model)
-
-        return LOSS    
-
-    def train_one_step(self):
-        self.model.train()
-        #step 1
-        X, y = self.get_next_test_batch()
-        self.optimizer.zero_grad()
-        output = self.model(X)
-        loss = self.loss(output, y)
-        loss.backward()
-        self.optimizer.step()
-            #step 2
-        X, y = self.get_next_train_batch()
-        self.optimizer.zero_grad()
-        output = self.model(X)
-        loss = self.loss(output, y)
-        loss.backward()
-        self.optimizer.step(beta=self.beta)
