+++ main.py
@@ -6,11 +6,10 @@ import argparse
 import importlib
 import random
 import os
-from fedl.servers.serveravg import FedAvg
-from fedl.servers.serverapfl import APFL
-from fedl.servers.serverpsnl import pFedMe
-from fedl.servers.serverperavg import PerAvg
-from fedl.trainmodel.models import *
+from FLAlgorithms.servers.serveravg import FedAvg
+from FLAlgorithms.servers.serverpFedMe import pFedMe
+from FLAlgorithms.servers.serverperavg import PerAvg
+from FLAlgorithms.trainmodel.models import *
 from utils.plot_utils import *
 import torch
 torch.manual_seed(0)
@@ -38,16 +37,13 @@ def main(dataset, algorithm, model, batch_size, learning_rate, beta, lamda, num_
 
         # select algorithm
         if(algorithm == "FedAvg"):
-            server = FedAvg(dataset,algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters, local_epochs, optimizer, numusers,i)
+            server = FedAvg(dataset, algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters, local_epochs, optimizer, numusers, i)
         
         if(algorithm == "pFedMe"):
-            server = pFedMe(dataset,algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters, local_epochs, optimizer, numusers,K,personal_learning_rate,i)
-
-        if(algorithm == "APFL"):
-            server = APFL(dataset,algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters, local_epochs, optimizer, numusers,i)
+            server = pFedMe(dataset, algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters, local_epochs, optimizer, numusers, K, personal_learning_rate, i)
 
         if(algorithm == "PerAvg"):
-            server = PerAvg(dataset,algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters, local_epochs, optimizer, numusers,i)
+            server = PerAvg(dataset, algorithm, model, batch_size, learning_rate, beta, lamda, num_glob_iters, local_epochs, optimizer, numusers, i)
 
         server.train()
         server.test()
@@ -65,12 +61,12 @@ if __name__ == "__main__":
     parser.add_argument("--model", type=str, default="dnn", choices=["dnn", "mclr", "cnn"])
     parser.add_argument("--batch_size", type=int, default=20)
     parser.add_argument("--learning_rate", type=float, default=0.005, help="Local learning rate")
-    parser.add_argument("--beta", type=float, default=1.0, help="Average moving parameter for pFedMe, or Learning rate of Per-FedAvg, or Mixmodel of APFL")
+    parser.add_argument("--beta", type=float, default=1.0, help="Average moving parameter for pFedMe, or Second learning rate of Per-FedAvg")
     parser.add_argument("--lamda", type=int, default=15, help="Regularization term")
     parser.add_argument("--num_global_iters", type=int, default=800)
     parser.add_argument("--local_epochs", type=int, default=20)
     parser.add_argument("--optimizer", type=str, default="SGD")
-    parser.add_argument("--algorithm", type=str, default="pFedMe",choices=["pFedMe", "PerAvg", "FedAvg", "APFL"]) 
+    parser.add_argument("--algorithm", type=str, default="pFedMe",choices=["pFedMe", "PerAvg", "FedAvg"]) 
     parser.add_argument("--numusers", type=int, default=20, help="Number of Users per round")
     parser.add_argument("--K", type=int, default=5, help="Computation steps")
     parser.add_argument("--personal_learning_rate", type=float, default=0.1, help="Persionalized learning rate to caculate theta aproximately using K steps")
